GN4 official close-out: Arm A both samplers FAIL but heun8 degrades (53% coverage) where euler4 dies (2.9%) -- loop amplifies injected bias, Self Forcing starts from 53% not zero; Arm B bottleneck map measured (1.468x confirmed per-module); Arm C PARITY OK -- blocking item 2 cleared, teacher product pipeline passes every quality gate

+++ experiments/p1_flow_head/score_gate_night4.py
@@ -56,7 +56,7 @@ def ecapa_curve(path, win=4.0, hop=2.0):
 def rms_halves(path):
     x, _ = sf.read(path, dtype="float32")
     h = len(x) // 2
-    db = lambda a: 20 * np.log10(np.sqrt((a**2).mean()) + 1e-9)  # noqa: E731
+    db = lambda a: float(20 * np.log10(np.sqrt((a**2).mean()) + 1e-9))  # noqa: E731
     return round(db(x[:h]), 1), round(db(x[h:]), 1)
